// api/seashell.schema.ts
import { z } from "zod";
import { SeashellSchema } from "../prisma/generated/zod/index.js";

// We don't include 'id' because that is auto-generated.
const CreateSeashellSchema = z.object({
  name: z.string().min(1).openapi({ example: "Queen Conch" }),
  species: z.string().min(1).openapi({ example: "Strombus gigas" }),
  description: z.string().optional(),
  rarity: z.enum(["common", "rare", "legendary"]).default("common"),
});

const IdParamSchema = z.object({
  id: z.coerce.number().int().positive().openapi({
    param: {
      name: 'id',
      in: 'path',
    },
    example: 42,
  }),
});

const ErrorSchema = z.object({
  error: z.string().openapi({ example: "An error occurred" }),
  details: z.array(z.string()).optional()
});

// Output schema generated by prisma
const SeashellResponseSchema = SeashellSchema;

// Create a TypeScript type from the schema so we can use it in the Service
export type CreateSeashellDto = z.infer<typeof CreateSeashellSchema>;

export {CreateSeashellSchema, 
  SeashellResponseSchema, 
  IdParamSchema,
  ErrorSchema}